refactored config
