+++ exporter/ydbexporter/factory_test.go
@@ -1,8 +1,12 @@
 package ydbexporter
 
 import (
+	"context"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/ydbexporter/internal/config"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 	"go.opentelemetry.io/collector/component/componenttest"
+	"go.opentelemetry.io/collector/exporter/exportertest"
 	"testing"
 )
 
@@ -16,46 +20,44 @@ func TestCreateDefaultConfig(t *testing.T) {
 }
 
 func TestFactory_CreateLogsExporter(t *testing.T) {
-	// TODO: Uncomment after understanding problem with connecting to ydb
-	//factory := NewFactory()
-	//params := exportertest.NewNopCreateSettings()
-	//cfg := config.WithDefaultConfig(func(c *config.Config) {
-	//	c.AuthType = authType
-	//})
-	//
-	//exporter, err := factory.CreateLogsExporter(context.Background(), params, cfg)
-	//
-	//require.NoError(t, err)
-	//require.NotNil(t, exporter)
-	//require.NoError(t, exporter.Shutdown(context.TODO()))
+	factory := NewFactory()
+	params := exportertest.NewNopCreateSettings()
+	cfg := config.WithDefaultConfig(func(c *config.Config) {
+		c.Endpoint = "grpc://localhost:2136"
+		c.AuthType = authType
+	})
+
+	exporter, err := factory.CreateLogsExporter(context.Background(), params, cfg)
+
+	require.NoError(t, err)
+	require.NotNil(t, exporter)
+	require.NoError(t, exporter.Shutdown(context.TODO()))
 }
 
 func TestFactory_CreateTracesExporter(t *testing.T) {
-	// TODO: Uncomment after understanding problem with connecting to ydb
-	//factory := NewFactory()
-	//params := exportertest.NewNopCreateSettings()
-	//cfg := config.WithDefaultConfig(func(c *config.Config) {
-	//	c.AuthType = authType
-	//})
-	//
-	//exporter, err := factory.CreateTracesExporter(context.Background(), params, cfg)
-	//
-	//require.NoError(t, err)
-	//require.NotNil(t, exporter)
-	//require.NoError(t, exporter.Shutdown(context.TODO()))
+	factory := NewFactory()
+	params := exportertest.NewNopCreateSettings()
+	cfg := config.WithDefaultConfig(func(c *config.Config) {
+		c.AuthType = authType
+	})
+
+	exporter, err := factory.CreateTracesExporter(context.Background(), params, cfg)
+
+	require.NoError(t, err)
+	require.NotNil(t, exporter)
+	require.NoError(t, exporter.Shutdown(context.TODO()))
 }
 
 func TestFactory_CreateMetricsExporter(t *testing.T) {
-	// TODO: Uncomment after understanding problem with connecting to ydb
-	//factory := NewFactory()
-	//params := exportertest.NewNopCreateSettings()
-	//cfg := config.WithDefaultConfig(func(c *config.Config) {
-	//	c.AuthType = authType
-	//})
-	//
-	//exporter, err := factory.CreateMetricsExporter(context.Background(), params, cfg)
-	//
-	//require.NoError(t, err)
-	//require.NotNil(t, exporter)
-	//require.NoError(t, exporter.Shutdown(context.TODO()))
+	factory := NewFactory()
+	params := exportertest.NewNopCreateSettings()
+	cfg := config.WithDefaultConfig(func(c *config.Config) {
+		c.AuthType = authType
+	})
+
+	exporter, err := factory.CreateMetricsExporter(context.Background(), params, cfg)
+
+	require.NoError(t, err)
+	require.NotNil(t, exporter)
+	require.NoError(t, exporter.Shutdown(context.TODO()))
 }
